requirements.txt
graphviz
pydot
keras
tensorflow-datasets
tensorflow-gpu==2.4.0
matplotlib
